src/main/java/com/chidi/moneycalculator/operation/OperationService.java
package com.chidi.moneycalculator.operation;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
@RequiredArgsConstructor
public class OperationService {

    private final OperationRepository operationRepository;

    public List<Operation> getAllOperations() {
        return operationRepository.findAll();
    }

    public Optional<Operation> getOperationById(Integer id) {
        return operationRepository.findById(id);
    }

    public Operation saveOperation(Operation operation) {
        return operationRepository.save(operation);
    }

    public void deleteOperation(Integer id) {
        operationRepository.deleteById(id);
    }
}
